requirements.txt
certifi==2018.1.18
cycler==0.10.0
kiwisolver==1.0.1
llvmlite==0.22.0
matplotlib==2.1.0
numba==0.37.0
numpy==1.14.1
numpy-quaternion==2018.5.31.16.8.42
pandas==0.22.0
pkg-resources==0.0.0
psutil==5.4.5
pyparsing==2.2.0
pyquaternion==0.9.2
python-dateutil==2.7.0
pytz==2018.3
scipy==1.0.0
six==1.11.0
tornado==5.0
vprof==0.37.4
